% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NIRspline.R
\name{nirSpline.default}
\alias{nirSpline.default}
\title{nirSpline.default}
\usage{
\method{nirSpline}{default}(wavelengths, absorbance, concentration,
  norder = 4)
}
\arguments{
\item{wavelengths}{}

\item{absorbance}{}

\item{concentration}{}

\item{norder}{}
}
\value{
List of two items
  \item{betaMatrix}{TODO}
  \item{fittedMatrix}{TODO}
}
\description{
Default action for the nirSpline function
}
\examples{
require(fda)
data(tecator)
with(tecator, nirSpline(seq(850, 1050, length = 100), training[,1:100],
                        training[,123:125]/100))
}
\references{
Ramsay, J. O. (2006) Functional Data Analysis. New York: Springer.

 Dias, R., Garcia, N., Ludwig, G. and Saraiva, M. (2014) Aggregated functional
 data model for Near-Infrared Spectroscopy calibration and prediction,
 Journal of Applied Statistics
}
\seealso{
\code{\link{tecator}}
}
\keyword{Aggregated}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}

