% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tecator.R
\docType{data}
\name{tecator}
\alias{tecator}
\title{tecator}
\format{a \code{list} of \code{matrix} objects.}
\source{
<http://lib.stat.cmu.edu/datasets/tecator>
}
\usage{
tecator
}
\description{
This data set is available from
<http://lib.stat.cmu.edu/datasets/tecator>. Please refer to the documentation
for usage and permissions. From the description, "Each line contains the 100
absorbances followed by the 22 principal components and finally the contents of
moisture, fat and protein." For this package, the dataset was transformed into a
list of matrices.
}
\author{
Hans Henrik Thodberg
}

